pi_requirements.txt
paramiko
scp
Pillow
picamera
